mailru (mrhash): drop support for go1.9
