+++ backend/mailru/mrhash/mrhash.go
@@ -7,6 +7,7 @@ package mrhash
 
 import (
 	"crypto/sha1"
+	"encoding"
 	"encoding/hex"
 	"errors"
 	"hash"
@@ -77,6 +78,17 @@ func (d *digest) Sum(b []byte) []byte {
 		panic(hashError)
 	}
 	return copy.Sum(b)
+}
+
+// cloneSHA1 clones state of SHA1 hash
+func cloneSHA1(orig hash.Hash) (clone hash.Hash, err error) {
+	state, err := orig.(encoding.BinaryMarshaler).MarshalBinary()
+	if err != nil {
+		return nil, err
+	}
+	clone = sha1.New()
+	err = clone.(encoding.BinaryUnmarshaler).UnmarshalBinary(state)
+	return
 }
 
 // Reset resets the Hash to its initial state.
